+++ uikit/project4/project4/ViewController.swift
@@ -6,85 +6,39 @@
 //
 
 import UIKit
-import WebKit
 
-class ViewController: UIViewController, WKNavigationDelegate {
-    var webView: WKWebView!
-    var progressView: UIProgressView!
-    var websites = ["apple.com", "hackingwithswift.com"]
-    
-    override func loadView() {
-        webView = WKWebView()
-        webView.navigationDelegate = self
-        view = webView
-    }
+class ViewController: UITableViewController {
+    var websites = [""]
 
     override func viewDidLoad() {
         super.viewDidLoad()
-        
-        navigationItem.rightBarButtonItem = UIBarButtonItem(title: "Open", style: .plain, target: self, action: #selector(openTapped))
-        
-        progressView = UIProgressView(progressViewStyle: .default)
-        progressView.sizeToFit()
-        let progressButton = UIBarButtonItem(customView: progressView)
-        
-        let spacer = UIBarButtonItem(barButtonSystemItem: .flexibleSpace, target: nil, action: nil)
-        let refresh = UIBarButtonItem(barButtonSystemItem: .refresh, target: webView, action: #selector(webView.reload))
-        let back = UIBarButtonItem(title: nil, image: UIImage(systemName: "chevron.backward"), target: webView, action: #selector(webView.goBack))
-        let forward = UIBarButtonItem(title: nil, image: UIImage(systemName: "chevron.forward"), target: webView, action: #selector(webView.goForward))
-        
-        toolbarItems = [back, forward, spacer, progressButton, spacer, refresh]
-        navigationController?.isToolbarHidden = false
-        
-        let url = URL(string: "https://" + websites[0])!
-        webView.load(URLRequest(url: url))
-        webView.allowsBackForwardNavigationGestures = true
-        webView.addObserver(self, forKeyPath: #keyPath(WKWebView.estimatedProgress), options: .new, context: nil)
+
+        title = "Visit The Internet"
+        navigationController?.navigationBar.prefersLargeTitles = true
+
+        websites = ["apple.com", "hackingwithswift.com"]
     }
 
-    @objc func openTapped() {
-        let ac = UIAlertController(title: "Open page…", message: nil, preferredStyle: .actionSheet)
-        for website in websites {
-            ac.addAction(UIAlertAction(title: website, style: .default, handler: openPage))
+    override func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
+            return websites.count
+    }
+    
+    override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
+            let cell = tableView.dequeueReusableCell(withIdentifier: "Website", for: indexPath)
+            
+            cell.textLabel?.text = websites[indexPath.row]
+            
+            return cell
+    }
+    
+    override func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
+        if let vc = storyboard?.instantiateViewController(withIdentifier: "WebView") as? WebViewController {
+            vc.website = websites[indexPath.row]
+            vc.websites = websites
+            
+            navigationController?.pushViewController(vc, animated: true)
         }
-        ac.addAction(UIAlertAction(title: "Cancel", style: .cancel))
-        ac.popoverPresentationController?.barButtonItem = self.navigationItem.rightBarButtonItem
-        present(ac, animated: true)
     }
     
-    func openPage(action: UIAlertAction) {
-        let url = URL(string: "https://" + action.title!)!
-        webView.load(URLRequest(url: url))
-    }
-    
-    func webView(_ webView: WKWebView, didFinish navigation: WKNavigation!) {
-        title = webView.title
-    }
-    
-    override func observeValue(forKeyPath keyPath: String?, of object: Any?, change: [NSKeyValueChangeKey : Any]?, context: UnsafeMutableRawPointer?) {
-            if keyPath == "estimatedProgress" {
-                progressView.progress = Float(webView.estimatedProgress)
-            }
-    }
-    
-    func webView(_ webView: WKWebView, decidePolicyFor navigationAction: WKNavigationAction, decisionHandler: @escaping (WKNavigationActionPolicy) -> Void) {
-        let url = navigationAction.request.url
-
-        if let host = url?.host {
-            for website in websites {
-                if host.contains(website) {
-                    decisionHandler(.allow)
-                    return
-                }
-            }
-        }
-        
-        let ac = UIAlertController(title: "Website not allowed", message: "Sorry, that website isn't on the approved list.", preferredStyle: .alert)
-        ac.addAction(UIAlertAction(title: "OK", style: .default))
-        present(ac, animated: true)
-
-        decisionHandler(.cancel)
-        
-    }
     
 }
