+++ uikit/project4/project4/WebViewController.swift
@@ -1,0 +1,79 @@
+//
+//  ViewController.swift
+//  project4
+//
+//  Created by Cassie Wallace on 10/10/22.
+//
+
+import UIKit
+import WebKit
+
+class WebViewController: UIViewController, WKNavigationDelegate {
+    var webView: WKWebView!
+    var progressView: UIProgressView!
+    var website = ""
+    var websites = [""]
+    
+    override func loadView() {
+        webView = WKWebView()
+        webView.navigationDelegate = self
+        view = webView
+    }
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        
+        progressView = UIProgressView(progressViewStyle: .default)
+        progressView.sizeToFit()
+        let progressButton = UIBarButtonItem(customView: progressView)
+        
+        let spacer = UIBarButtonItem(barButtonSystemItem: .flexibleSpace, target: nil, action: nil)
+        let refresh = UIBarButtonItem(barButtonSystemItem: .refresh, target: webView, action: #selector(webView.reload))
+        let back = UIBarButtonItem(title: nil, image: UIImage(systemName: "chevron.backward"), target: webView, action: #selector(webView.goBack))
+        let forward = UIBarButtonItem(title: nil, image: UIImage(systemName: "chevron.forward"), target: webView, action: #selector(webView.goForward))
+        
+        toolbarItems = [back, forward, spacer, progressButton, spacer, refresh]
+        navigationController?.isToolbarHidden = false
+        
+        let url = URL(string: "https://" + website)!
+        webView.load(URLRequest(url: url))
+        webView.allowsBackForwardNavigationGestures = true
+        webView.addObserver(self, forKeyPath: #keyPath(WKWebView.estimatedProgress), options: .new, context: nil)
+    }
+    
+    func openPage(action: UIAlertAction) {
+        let url = URL(string: "https://" + action.title!)!
+        webView.load(URLRequest(url: url))
+    }
+    
+    func webView(_ webView: WKWebView, didFinish navigation: WKNavigation!) {
+        title = webView.title
+    }
+    
+    override func observeValue(forKeyPath keyPath: String?, of object: Any?, change: [NSKeyValueChangeKey : Any]?, context: UnsafeMutableRawPointer?) {
+            if keyPath == "estimatedProgress" {
+                progressView.progress = Float(webView.estimatedProgress)
+            }
+    }
+    
+    func webView(_ webView: WKWebView, decidePolicyFor navigationAction: WKNavigationAction, decisionHandler: @escaping (WKNavigationActionPolicy) -> Void) {
+        let url = navigationAction.request.url
+    
+        if let host = url?.host {
+            for website in websites {
+                if host.contains(website) {
+                    decisionHandler(.allow)
+                    return
+                }
+            }
+        }
+    
+        let ac = UIAlertController(title: "Website not allowed", message: "Sorry, that website isn't on the approved list.", preferredStyle: .alert)
+        ac.addAction(UIAlertAction(title: "OK", style: .default))
+        present(ac, animated: true)
+    
+        decisionHandler(.cancel)
+    
+    }
+    
+}
